Set up HTML table

+++ Models/AveragedSpreadspoke.cs
@@ -8,14 +8,14 @@ schedule_date	schedule_season	schedule_week	schedule_playoff	team_home	score_hom
 public class AveragedSpreadspoke
 {
     public int ScheduleSeason { get; set; }
-    public int SpreadFavorite { get; set;}
+    public float SpreadFavorite { get; set;}
 
-    public int SpreadActual { get; set;}
+    public float SpreadActual { get; set;}
 
-    public int SpreadDifference {get; set;}
+    public float SpreadDifference {get; set;}
 
     //Using int so I can average the number of games with SpreadCorrect by Year
-    public int SpreadCorrect { get; set;}
+    public float SpreadCorrect { get; set;}
     public AveragedSpreadspoke()
     {
